+++ python/morpheus/morpheus/_lib/include/morpheus/stages/monitor.hpp
@@ -58,10 +58,10 @@ class MORPHEUS_EXPORT MonitorStage : public mrc::pymrc::PythonNode<std::shared_p
      * @param determine_count_fn : A function that computes the count for each incoming message
      */
     MonitorStage(const std::string& description,
-                 const std::string& unit                                           = "messages",
-                 indicators::Color                                                 = indicators::Color::cyan,
-                 indicators::FontStyle font_style                                  = indicators::FontStyle::bold,
-                 std::optional<std::function<int(sink_type_t)>> determine_count_fn = std::nullopt);
+                 const std::string& unit                                              = "messages",
+                 indicators::Color                                                    = indicators::Color::cyan,
+                 indicators::FontStyle font_style                                     = indicators::FontStyle::bold,
+                 std::optional<std::function<size_t(sink_type_t)>> determine_count_fn = std::nullopt);
 
   private:
     subscribe_fn_t build_operator();
@@ -74,7 +74,7 @@ MonitorStage<MessageT>::MonitorStage(const std::string& description,
                                      const std::string& unit,
                                      indicators::Color text_color,
                                      indicators::FontStyle font_style,
-                                     std::optional<std::function<int(sink_type_t)>> determine_count_fn) :
+                                     std::optional<std::function<size_t(sink_type_t)>> determine_count_fn) :
   base_t(base_t::op_factory_from_sub_fn(build_operator())),
   m_monitor_controller(MonitorController<sink_type_t>(description, unit, text_color, font_style, determine_count_fn))
 {}
@@ -124,7 +124,7 @@ struct MORPHEUS_EXPORT MonitorStageInterfaceProxy
         const std::string& unit,
         indicators::Color color          = indicators::Color::cyan,
         indicators::FontStyle font_style = indicators::FontStyle::bold,
-        std::optional<std::function<int(typename MonitorStage<MessageT>::sink_type_t)>> determine_count_fn =
+        std::optional<std::function<size_t(typename MonitorStage<MessageT>::sink_type_t)>> determine_count_fn =
             std::nullopt);
 };
 
@@ -136,7 +136,7 @@ std::shared_ptr<mrc::segment::Object<MonitorStage<MessageT>>> MonitorStageInterf
     const std::string& unit,
     indicators::Color text_color,
     indicators::FontStyle font_style,
-    std::optional<std::function<int(typename MonitorStage<MessageT>::sink_type_t)>> determine_count_fn)
+    std::optional<std::function<size_t(typename MonitorStage<MessageT>::sink_type_t)>> determine_count_fn)
 {
     auto stage = builder.construct_object<MonitorStage<MessageT>>(
         name, description, unit, text_color, font_style, determine_count_fn);
